refactor: 자료실 공용 DataTable 적용
